Removed spaces vs. %20 handling which shouldn't be necessary if one properly works with URL, URI and File.
Utils.formatPath might be a bit special as it seems to be cosmetic only, but why is not dealing with german umlauts and such? Why is it reducing ALL spaces to ONE %20? How can one be sure the string is not used as some relative path if it is part of a path in the caller?

Looks to me removing that line to get rid of %20 handling altogether is not that bad...